requirements.txt
flask-restplus
werkzeug==0.16.1
pymongo
requests
dnspython
pylint
coverage
flask_cors